pyproject.toml
[project]
name = "course_work_1"
version = "0.1.0"
description = ""
authors = [
    { name = "Лусине", email = "lumahinyan@mail.ru" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"
pytest = "^8.3.5"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.2"
mypy = "^1.15.0"
black = "^25.1.0"
isort = "^6.0.1"

[tool.black]
line_length = 119
exclude = ".git"

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = "venv"

